LayoutTests/platform/mac-catalina/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-combining-expected.txt


FAIL Spacing around ≠ assert_approx_equals: expected 77.75 +/- 1 but got 67.5
FAIL Spacing around =⃒ assert_approx_equals: expected 77.75 +/- 1 but got 53.90625
FAIL Spacing around |̸ assert_approx_equals: expected 61.09375 +/- 1 but got 68.546875
FAIL Spacing around |⃒ assert_approx_equals: expected 61.09375 +/- 1 but got 41.859375
 
=
 

 
≠
 

 
=⃒
 

 
|
 

 
|̸
 

 
|⃒
 
